Migrating cache to be a templated type
Also starting to add some material pieces but I should add a color type
before doing that, along with the differential/shading geometry

The projection on #2 is still wrong too, I'm not sure what's going on

+++ src/render/camera.cpp
@@ -21,7 +21,7 @@ Camera::Camera(const Transform &cam_world, float fov, int xres, int yres) : cam_
 		screen[1] = 1 / aspect_ratio;
 	}
 	std::cout << cam_world << std::endl;
-	Transform screen_raster = Transform::scale(static_cast<float>(xres), static_cast<float>(yres), 1)
+	Transform screen_raster = Transform::scale(xres, yres, 1)
 		* Transform::scale(1.f / (2 * screen[0]), 1.f / (2 * screen[1]), 1)
 		* Transform::translate(Vector{screen[0], screen[1], 0});
 	Transform raster_screen = screen_raster.inverse();
